refactor: 이미지를 출력하는 부분 컴포넌트로 분리
